accessibility additions + non-automatic music
Users must click 'start music' button in order to get audio to play--will be less jarring. Added necessary labeling compliant to the WAVE accessibility tool, with all errors/warnings cleared with the exception of a warning for my video element. not much to be done, considering video content will be dynamic and can't be captioned.